+++ test/IntegrationTests/gbathree.cpp
@@ -87,7 +87,7 @@ TEST_CASE("Gbathree") {
 
     REQUIRE(3 == array.size());
 
-    for (int i = 0; i < 3; i++) {
+    for (size_t i = 0; i < 3; i++) {
       REQUIRE(50 == array[i]);
     }
   }
@@ -112,11 +112,11 @@ TEST_CASE("Gbathree") {
     REQUIRE(array.success());
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       JsonArray& nestedArray = array[i];
       REQUIRE(4 == nestedArray.size());
 
-      for (int j = 0; j < 4; j++) {
+      for (size_t j = 0; j < 4; j++) {
         REQUIRE(34 == nestedArray[j]);
       }
     }
@@ -130,7 +130,7 @@ TEST_CASE("Gbathree") {
 
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       REQUIRE(2 == array[i]);
     }
   }
@@ -143,7 +143,7 @@ TEST_CASE("Gbathree") {
 
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       REQUIRE(2 == array[i]);
     }
   }
@@ -155,12 +155,12 @@ TEST_CASE("Gbathree") {
     REQUIRE(array.success());
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       JsonArray& nestedArray = array[i];
 
       REQUIRE(4 == nestedArray.size());
 
-      for (int j = 0; j < 4; j++) {
+      for (size_t j = 0; j < 4; j++) {
         REQUIRE(15 == nestedArray[j]);
       }
     }
@@ -173,11 +173,11 @@ TEST_CASE("Gbathree") {
     REQUIRE(array.success());
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       JsonArray& nestedArray = array[i];
       REQUIRE(4 == nestedArray.size());
 
-      for (int j = 0; j < 4; j++) {
+      for (size_t j = 0; j < 4; j++) {
         REQUIRE(15 == nestedArray[j]);
       }
     }
@@ -191,7 +191,7 @@ TEST_CASE("Gbathree") {
 
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       REQUIRE(2 == array[i]);
     }
   }
@@ -204,7 +204,7 @@ TEST_CASE("Gbathree") {
 
     REQUIRE(4 == array.size());
 
-    for (int i = 0; i < 4; i++) {
+    for (size_t i = 0; i < 4; i++) {
       REQUIRE(2 == array[i]);
     }
   }
